requirements.txt
numpy
matplotlib
pillow
scikit-learn
scikit-image